+++ ali/common/src/main/java/com/yanny/ali/api/IServerUtils.java
@@ -68,6 +68,9 @@ public interface IServerUtils extends ICoreServerUtils<IServerUtils>, ICommonUti
     <T extends LootItemFunction> ItemStack applyItemStackModifier(IServerUtils utils, T function, ItemStack itemStack);
 
     @NotNull
+    TooltipBuilder getEnumTranslation(IServerUtils utils, Enum<?> value);
+
+    @NotNull
     RangeValue convertNumber(IServerUtils utils, @Nullable NumberProvider numberProvider);
 
     @Nullable